// apps/sales-copilot/src/generated/models/Crf5c_account1sModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crf5c_account1scrf5c_creditstatus = {
  995340000: '__',
  995340001: '__',
  995340002: '__'
} as const;
export type Crf5c_account1scrf5c_creditstatus = keyof typeof Crf5c_account1scrf5c_creditstatus;
export const Crf5c_account1scrf5c_paymentstatus = {
  995340000: '__',
  995340001: '__',
  995340002: '___'
} as const;
export type Crf5c_account1scrf5c_paymentstatus = keyof typeof Crf5c_account1scrf5c_paymentstatus;
export const Crf5c_account1scrf5c_region = {
  995340000: '__',
  995340001: '__',
  995340002: '__',
  995340003: '__'
} as const;
export type Crf5c_account1scrf5c_region = keyof typeof Crf5c_account1scrf5c_region;
export const Crf5c_account1scrf5c_tier = {
  995340000: 'S',
  995340001: 'A',
  995340002: 'B',
  995340003: 'C'
} as const;
export type Crf5c_account1scrf5c_tier = keyof typeof Crf5c_account1scrf5c_tier;
export const Crf5c_account1sstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crf5c_account1sstatecode = keyof typeof Crf5c_account1sstatecode;
export const Crf5c_account1sstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crf5c_account1sstatuscode = keyof typeof Crf5c_account1sstatuscode;

export interface Crf5c_account1sBase {
  crf5c_account1id: string;
  crf5c_address?: string;
  crf5c_creditstatus?: Crf5c_account1scrf5c_creditstatus;
  crf5c_email?: string;
  crf5c_industry?: string;
  crf5c_lastcontactedon?: string;
  crf5c_lastinteractiondate?: string;
  crf5c_latitude?: string;
  crf5c_longitude?: string;
  crf5c_name: string;
  crf5c_notes?: string;
  crf5c_ownerid: string;
  crf5c_paymentstatus?: Crf5c_account1scrf5c_paymentstatus;
  crf5c_phone?: string;
  crf5c_region?: Crf5c_account1scrf5c_region;
  crf5c_tier?: Crf5c_account1scrf5c_tier;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Crf5c_account1sstatecode;
  statuscode?: Crf5c_account1sstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Crf5c_account1s extends Crf5c_account1sBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  crf5c_creditstatusname?: string;
  crf5c_paymentstatusname?: string;
  crf5c_regionname?: string;
  crf5c_tiername?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}